+++ src/tnbbeta_vae/models/losses/__init__.py
@@ -1,1 +1,5 @@
 """ELBO and other loss terms for VAE training."""
+
+from tnbbeta_vae.models.losses.elbo import monte_carlo_elbo
+
+__all__ = ["monte_carlo_elbo"]